Don't increment program counter before insn done..
This is leading to funkyness. It turns out the whole of the rest of
symex related stuff is geared towards fetching details on the currently
executing piece of code from state.source.pc -- which is fine. Just, it
puts us in a bind with regard to working out when to context switch, and
what program counter to have afterwards.

So instead I'm about to head down the route of splitting the decision to
make a context switch, and actually doing it. This should neatly step
around the problem of cloning an execution state in mid execution.