Attempt to split up art state duplication code.
So the plan is to have code decide whether to generate a context switch
during symbolic execution, but to actually perform the context switch
_after_ symex has completed the step. This steps around problems
involving duplicating states in the middle of execution.

Unfortunately it doesn't quite work yet - there's quite a bit of trouble
with how context switch points that aren't generated by code (i.e., the
end of a thread or likewise) interface with this.

Plus, now all the names are wrong.