Simplify state guard before using
Vastly improves performance in terms of number of interleaves generated;
02_sync01 has been running 40 minutes now, and with this completes in
30s.

The number of interleaves generated are vastly decreased this way, but
I'm not certain it's as few as they have been; perhaps it's because
constant propagation was in play before?

(This makes 01_malloc_11 fail again. Weird - but then again it turns out
 to be suffering a conversion error, not an ART error, so it's fine)