Relax funcptr pointer checking
Right now when derefing a function pointer, because we've broken the
function pointer tracking when sharing between threads, one ends up with
no targets of a dereference, which means CBMC generates an
invalid object to represent the dereference. That then gets stuffed into
an assertion, which always fails, regardless of what happens at runtime.

So instead, don't analyse function pointers, instead assert that, if
it's a pointer, that it's a valid one.