Replace strabs name seperator from # to \\
This because since making changes to how renaming works, we've been
taking '#' charactesr in a symbol name as being part of syntax (i.e.,
I assume that anything with a '#' in its name  is a l2renamed name,
which is safe because you can't do that from C).

However, of course string abstraction also uses '#' in symbol names for
the same reason. Which royally messes with the l2 renaming.

Clearly the l2 renaming needs refactoring in the future (with, y'know,
types), but til then the faster fix is to drop the use of '#' characters
in strabs names.