Revert "Exchange start/end data through arrays"
(But keep end data pumped through arrays).

So, all the passing start and end data through arrays is all very well.
However, the pointer tracking for arrays of pointers just congeals into
one set of possibilities: so for an array of function pointers, a read
from one element of the array could be any of the values put at any
index, according to the pointer tracking (and that's a fine
approximation).

Unfortunately that approximation means that the creation of threads with
potentially different function pointers will result in _all_ thread
functions being run in this thread, with a thread guard determining
which one actually gets evaluated. This is _correct_, but
_vastly_inefficient_. So, keep thread start data pumped through internal
ESBMC gubbins to eliminate this unpleasentry.

Then net upshot of this is that creation of a pthread is still
explicit-state dependant; which is fine, because I don't think anything
we have right now requires nondeterministic thread creation (and it
wouldn't have worked before anyway).

However the end value of a thread still has its data pumped through
arrays and foo, so we can nondeterministically join a thread.

This reverts commit 3f44c7f4eb41e9c014223a6189080785e466724e.

Conflicts:

	ansi-c/library/pthread_lib.c