Improve 00_prodcons efficiency
To make this run in less than half an hour, it's run --round-robin.
However, due to some changes in context-switching behaviour, it no
longer works like that. So, fix it up so that it works like it did; by
stopping the generating thread from context switching until all the
other threads are created.

(The previous commit was related to this as well.)