Add thread library hooks for main thread start/end
The point of these is so that the threading library gets notified when
the main thread starts and ends, so that deadlock detection actually
works properly (currently the main thread quitting doesn't register as
an end of thread). IMO the internal model checker goo can be 100%
ignorant about this going on beneath its feet, and that's the way it
should be.