Reduce context switches in mutex_lock_check
So, we don't need multiple context switches in this. Locking something
is atomic in itself. For deadlock, we'll always be able to construe an
interleaving where all other threads have exited or are also blocked
before this thread tries to lock, leading to the deadlock assertion. The
test 01_cond_05 shows that deadlock is still detectable.

Question: what here, and before this patch, was ensuring that the state
guard became false if the lock was already locked before this
interleaving/function ran?