Don't exit on thread guard being false
For --interactive-ileaves

This actually happens quite often; there's some portion of code that's
run that causes a context switch, in an if/else branch that has a false
guard. And the rest of the trace is valid, just not that bit.

Still, print a warning each time so that you can trace back something if
assume(false) occurs.