Move not-deadlocked assume out of atomic block
The desired behaviour here is that we can context switch to some other
thread and do something things, potentially ending up discovering a
global deadlock. But we also want to stop exploration from all points
after this for efficiency.

So have an assume that we're not deadlocked to block further
exploration, but put it _after_ a context switch point so that other
threads can advance and possibly discover global deadlock.