Context bound a successful test
I don't enjoy bounding tests that don't have a specific counter example
they're looking for - I could just as easily set the bound to 1 and not
detect a fault, but then potential errors in the future wouldn't be
picked up.

So, I've picked 3 here, because 00_race16 is a very similar test that
_does_ find a counterexample with 3 context switches, so I figure this
one should be able to do that too, if it had an error to find.