Only store decl history for a particular frame.
I think what this does is ensure that redeclarations of variables such
as:

for (i = 0; i < n; i++) {
  int foo;
  assert(somestuff);
}

(Where foo is repeatedly redeclared)... ensures that places like this
always get a fresh new variable each time we run into a decl. Which is
enforced if it turns out we've already seen it declared.

There's nothing in that, afaik, that requires us to have that
declaration history across all of a thread state, so just localise it to
a particular function.