Track the current thread state being worked on
Stored in goto_symext so that we don't have to keep on passing around a
reference to whatever state we're working on. There are no circumstances
where symex operations are applied to more than one state at a time, and
if there were they would have to be handled in execution_statet anyway.