Begin implementing type thrashing arrangement.
The plan: instead of trying to rename a bunch of type things at runtime,
which as it turns out in various circumstances, can lead to a _lot_ of
memory being used up...

Instead of that, we instead try to eliminate all symbolic types that we
can out of the program generated before dumping it into goto functions.
After all, the compiler _has_ to be able to do this, so so can we.

Start off by collecting type information prior to actually modifying
things.