Handle differently-scoped type definitions
Turns out that structs can be defined local to functions or other...
things. So "c::tag-" doesn't pick up all struct definitions, only
globally scoped ones.

Instead, search for "tag-" in a symbol name - if it's in there, it's a
struct definition. The '-' character can't creep in any other way.

Oh for types.