Fiddle with equality type assertion
We can have a renamed pointer and an unrenamed pointer turn up here,
resulting in an assertion failure. However, even though the contents of
the type irep differ, if they're both pointers then it's a safe
conversion. So, keep the type assertion, but allow both operands being
pointers to fufil the assertion.