Munge monitor insn ordering.
It turns out that for all of the past, we've been updating the boolean
propositions value with the _previous_ value of the equality. Because
the instruction is inserted before the assignment that triggered it!

So, move it into the body of the atomic block, update the proposition,
and also switch to the monitor in that block. This shouldn't affect any
reachability thing because switching to the monitor isn't a thread
exploration thing any more. And the atomic_end is still going to trigger
a context switch that would have been caused by the original assignment.