Remove "no mapping" error functionality.
New plan: mapping class contains method pointers of some object; and
when mapping we call said method pointers. This means that we get to
have the code implementing a mapped function _inside_ a class, which is
for the plus plus simplicity. However this means we need the generating
class to provide a suitable default method.