Define argument to convert_smt method.
The upshot of all this wrangling is this: the types that are passed
through the expr2t class have to be either extremely simple, or a void
pointer, to allow for code on either side to share data between each
other. An immense amount of templating would be absolutely horrific.

So instead we take the solution of passing a void pointer through
methods in expr2t types. This isn't type safe, which may make C++
enthusiasts upset, however it works, is efficient, and doesn't involve
templating. It also involves ultimately only two casts.

Additionally, this mechanism is 100% safe if we keep to the accepted way
of using this kind of mechanism, passing your own object into the
convert / whatever method, and a handle to whatever data you're going to
be operating with as the argument.