New tack again: templated derived classes.
So I thought that this'd be impossible, passing down a derived typename
to a template class that we're extending. I figured it'd be insane, and
that due to instanciation/typing order that the compiler would chuck it
out.

However, read http://www.hackcraft.net/cpp/templateInheritance/
and the "Templates Parameterised by a Derived Class" section.

It turns out that this really is a thing (and it's somewhat sane), that
you can template a class using your own new type. So long as the
template class never needs to know the layout of the derived class,
which is 100% reasonable.

C++ has me suprised. That it hasn't exploded in my face yet.