Fix non-trivial inheretance in templated classes.
If we have:

signedbv_type2t -> bv_type2t -> type2t

Then we need to insert type_body between each inheretance mark (->).
However, that means we need to modify the type_body template to inherit
from the right place.

This is solved by the lump of specialization in this patch.