Don't cause an error when we have a missing conv
So the current situation is that there's an array converted in
02_eureka21 that has '+' as part of its size (which doesn't get
simplified). That plus also means it doesn't get converted, which means
that the whole irep doesn't get converted.

This shows up the fact that we're going to end up with some pieces of ireps
that depend upon several ireps, which depend upon other ireps, which
form a chain. So we can either convert them all at the same time
(messy), or live with temporary conversion errors.

So instead, don't convert error if we don't find an irep. Instead, the
plan should be to slowly move them all over, then start removing the
deprecated conversion methods until we find an edge case that still
isn't being converted. This is the only good way to end up with a
conversion situation that I can git-bisect.