Constify convert_smt_type methods.
This is because they're also going to be used from things like
z3_get.cpp, where we're handed a const prop_conv object to work with.

We can cope with having const type converter functions (ish), happily
the irep conversion functions don't need to be const. But ultimately,
this problem stems from code elsewhere wishing to be able to do withis
with a const prop_conv object.