Dump current debug status of type munging
Currently trying to fix the cpp type renaming situation; the current
piece of code identifies types by seeing that their containing
"named_irep"s name is type or subtype.

This seems to fix some things, but not all. The initial renaming of all
types seems to be broken right now.