Add storage for modelling-data array names
So in fixing the CPP pointer modelling situation, I've discovered that
there's absolutely no way for c::__ESBMC_alloc to be linked into C++
code and be initialized correctly; we can either have it in C code, or
have cpp::__ESBMC_alloc instead, but we can't cross them over.

This is because you can't declare _anything_ in C++ without it having
the cpp namespace; you can declare it extern C but that won't give it an
initializer; and all attempts of mine to suck c::__ESBMC_alloc out of
the C library with a zero initializer have failed.

So, I'm going to make the symex parts of ESBMC determine which modelling
array they should be using once they start symexing, and stick with
that.

This is all fixable by making alloc/dealloc functions part of the C
library, and not stitching together manipulations of these arrays at
symex time.