Eliminate a superfluous context
I have no idea why this additional context exists, and it isn't
documented in CBMC anywhere. From how it's used, it would appear to be
something to do with being able to look up symbols in multiple
namespaces?

Either way, this is making a tiny test program throw a segfault at the
ns.lookup in goto_symext::goto_symext. Again, no idea why, but this
fixes it.