Invert regression test outcome
When I run this test using 0d7eb963, it reports an invalid pointer
during initialization of some objects. When run --no-slice, this goes
away.

With the new arrangement (this commit), it always reports successful
verification. I can't see the error here, so I'd suggest it's because of
the previous stuff that's been fixed.