Allow recursive type definitions
This occurs, as documented in the comment, when there's a method in a
class that takes an object argument that is itself. Because method
definitions are in types, that means that a type can have a self
reference in a method definition.

The current approach is to let this slide and hope that that information
isn't used anywhere during the symbolic execution, which I don't believe
it is.