Support converting new arrays
This introduces a functional change; arrays were being created with
names it seems, taken from the tag in the array. This is extremely
likely to be causing bugs, as seein the past with struct, where two
constant arrays with different contents that get the same name here
cause the formula to be unsat.

The fix is to use Z3_mk_fresh_const, which should stop that from
happening.