Attempt to support array_of conversions.
This attempts to keep the same behaviour as before, but I've commented
out a couple of places, arrays of structs/unions, that I don't think are
likely to exist and thus likely to need maintenence right now.

Unfortunately the blasted thing suffers once more from the use of names
rather than fresh constants.