Work around a counterexample extraction problem.
This is some pointer logic thing that I don't understand, but that
doesn't expect pointers to have negative offsets from their objects,
which is valid. (Well, valid in that that's probably why we're
generating a counterexample).

Disable this assertion, add warning, we'll have to worry about it in the
future.