Force all logical ops to have bool type.
This is handled by having the constructor for the base logical operation
class to allocate a new bool object, so there's no scope for the expr
not being a bool type.

This is another situation where we eventually need to worry about
the memory allocation; ideally it's going to come out of some global
bool pool.