Install the int2bv trainwreck.
int2bv acts as an uninterpreted function. That means that here, we're
going to end up with almost random data coming out of these conversion
routines. And this is exactly how the current conversion function works!

The reason for this appears to be that there's no good detection of
bit operations in the produced program code, and thus such expressions
can turn up at our door without us being forced into bit vector mode.
Which sucks. The correct fix is to improve the frontend.

I've copied this trainwreck into the new irep situation because it's
supposed to only be a storage refactor, not a functionality refactor.