Make expr flattener work with subtracts.
This makes ireps with multiple operands (i.e., 3 add operands or
whatever) left recursive? Or whatever the case; the  point is four adds
become from:

  +  a  b  c  d
  +  a  b
  +     ^  c
  +        ^  d

If you get the drift; and the same for subs. By doing it this way, we
can ensure subs are converted correctly. The first most operand is the
original value from which everything else is subtracted from.

It's late, some of that might make sense.