Implement mul conversion.
This is pretty simple in a couple of cases - in int mode, we just
convert, and multiply. For integer bv's, that's bvmul. For fixedbv
multiplies in non-integer mode, we encounter... some peculiar code. I
don't know how it works, it's probably wrong, but it's copied from the
current convert_mul.

This can be handled by some future set of test cases which should
actually be testing these combinations of things.