Implement conversion for with exprs
Includes a trap to see whether or not struct/union with's are used, and
how. I can't currently work out what's going on with the
convert_member_name method, and how I can extract the same data out of
index expressions on migration.

Looking like that code path isn't going to be triggered though...