Hack around Z3s annoying returning-a-sym behaviour
When returning a model, for some weird reason z3 likes to stick in
symbols with the right type, but no further operands.

Now, these symbols might resolve to more expressions. But, it's not
immediately possible to extract from these the relevant pointer details.
Which is unpleasent. Hack around this for the moment, we can fix it up
in the future.