Correct implementation of invalid-object.
Turns out I got it entirely wrong; something somewhere else was expecing
this irep to become the representation of the invalid object, and all we
have to do is return a symbol for it.

I wonder whether in the future there'll be any kind of worry about
whether or not the pointer offset in the invalid object symbol can be
abused. However, so long as it's only ever used in single, indepenant
assertions, I think it's fine.