Make pointer offs size calculation less accurate.
It would appear (00_verisec_NetBSD-libc__anyMeta_int_ok) that numerous
things go wrong calculating various type sizes (see the previous
commit). Also note that it's believed in esbmc that the size of A is
believed to be one element; CBMC parses it as three; and just generally
the calculation is wrong there, if you look at the output of the
preprocessing.