Don't make unrecognized exprs abort now.
Almost at the end; the only problem visible right now is that
"dynamic_object" keeps on turning up at the solver, there isn't a
current implementation for it, and so things explode. In the past this
became a conv_error, so let it become a conv_error now.

I'm going to start stripping out code.