Remove some features from conv_error class.
There's now no longer an irep at each point to be printed, so get rid of
that. One might say that we should be printing the new irep, but the
real plan is to fix all conv_errors so that they're no longer
conv_errors (or are fatal).

Also get rid of the stack trace. That's never helped anyone
unfortunately, on account of for some reason it not printing symbols ;_;