Start using convert_bv everywhere.
This is more elaborate than it seems: the expression passed into
convert_bv may be cached, forever, and thus it's not valid to pass a
stack-allocated piece of irep to convert_bv.

Some of this patch fixes a few of these things, but segfaults are
occuring, so I figure it's not over yet.