Macro up some casting in Z3 conv.
This is the cast from a void pointer type (used to make irep2 agnostic
about what data is being shuttled through it) back to the original type,
which is a z3 ast reference. Now, this used to issue me with a bunch of
"type puned pointers are going to make you cry" warnings, and they went
away for some reason. This should satisfy absolutely everything by using
a reinterpret cast, which should be enough warning to the compiler
that things are risking aliasing.