Remove a wedge of conversion special casing.
I expect that the reason for these existing was that when CBMC used
only SAT, everything had to be expressed in bools, so it was better to
convert bool stuff immediately, then allow convert_rest to remunge
everything else into bools.

However this isn't the case now, and the backend solver is entirely
capable of converting anything immediately. In fact, removing this code
improves regression test runs from 8:20s to 8:00s immediately.

Plus, this means we can start removing all of the prop-specific stuff,
and just leave it to the solver converter.