Start routines for migrating /back/ to old irep.
Unfortunately this is necessary -- we need to use from_expr to print out
C code for counterexamples and other debugging goo, and it's not viable
to start rewriting parts of the C parser or otherwise messing with it.

So, the C and C++ frontend have to take an old exprt, so migrate new
exprs back to old ones.