Add default initializer for expr_body2 arg.
So this is an extremely advanced situation. In the case of a field not
being used (template arg gets blank_value), the constructor arg still
has to exist, but it can get initialized using some default value and
ignored (and no data stored at runtime).

This requires that there's some default value for the blank_value member
template thingy. It would also apparently require a default value for
everything else that can be used as a member type.

However, it looks like that default value isn't used, /or even
instanciated/, unless the caller doens't provide that argument. So, we
should be able to get away with setting a default argument of
::defaultval for all arguments, and it'll only get instanciated (and
produce an error) if the caller doesn't provide an argument for a _real_
member template thingy.