Throw one million shoehorns at expr_body2.
So this makes expr_body2 inherit from field2's fieldtype. And it
provides a constructor argument for it, and it provides a default value
for it.

But, that default value has to behave weirdly for the empty name type
because... well, it's too late and I'm really to tired to explain. But
it requires the third constructor for name_class_empty.