Define some infrastructure for pretty printing.
So the plan is this: in expr_body2::tostring we call a series of
tostring functions through the field templates. Because those types are
superclasses of the expr2_body2 template, this should lead to a generic
ability to select a particular member that's been included in this
type/template/whatever.

Then all we need to do is from there call type_to_string, where we can
stuff template specialisations for each particular type we need to
print. Hurrah.