Demonstrate cmp templating situation works.
So, we access our local piece of type generic data in the same way as
with tostring, by calling field1::fieldtype::whatever. And we get the
corresponding value out of the class we're comparing against via means
of a static cast to the member tag we're dealing with. This works
because the member tag is unique, and our corresponding value inherits
from it.

Required a slight patch to the current member tag cmp implementation,
which is just fine.