Remove struct_union_type2t
This involves a number of backflips. Primarily is the realisation that
we need to have /some/ kind of accessor or ability to treat types and
unions generically. So, we add some accessing functions to fetch a few
references, and then spread them around everywhere.

And then, because struct_union_type2t can't be used for convert_smt_type
despatching, we have to hack some stuff; convert_struct_union_type for
example.

It's disapointing that we've lost the ability to control whole swathes
of irep on the type despatching front; but that's too bad, I think the
current templating goo is going to make everything much more
maintainable.