Run with nonatomic boost operations.
ESBMC doesn't use multithreading right now; and it probably won't for a
long time. So, there's no point issuing atomic reference count
operations that are just going to slow us down.

On 12_lms_new this causes somewhere between 2 to 10ms of speed
improvement (dodgy, system activty at the time), so it's not a speed
decrease.