Start working on a more specific irep container.
That is, a version of shared_ptr that will contain a more specifically
typed pointer. The idea is to replicate the current situation that
Kroening has developed; where a non-const access to a container causes
it to detach from the existing object, clone a new one, and store that.

After some thought this is actually quite a nice concept, allowing one
to not worry too much about either memory management or intefering with
other users of your data.

However it's certainly not the most efficient; in a glorious irep3
future, we should be able to eliminate the need to edit code in place,
and instead associate portions of code with the data they're going to
operate on.

I.E., you have a piece of code with "symbolic symbols", that get linked
up to real symbols in some kind of lookup table. Same situation for
dereferences. Then symbolic execution would just be associating portions
of code with renamed symbols and pointer data.

But I digress; this class is just going to cast-down a pointer, and
hopefully perform detach action when non-const operations occur.