Rotate a little over container get/operator->
So, C++ can't have return type overloading. Sad, but probably necessary.
This means that we can't make the container decide whether or not a
const get()/operator-> operation is const or not. So, we'd end up
detatching all the time.

To get around this, change the semantics a little; operator-> now
returns a const pointer, and get() a volatile one. Dunno how irept does
it; but never mind.