Change how simplify works slightly.
So, we can avoid an unnecessary function call by making the
simplification routines part of the expr class. Which is good. It also
means that we break what I said in the previous commit about keeping
storage and implementation seperate, but meh, what I really care about
is organization. (I.E., it isn't all in the same file).