Shiver my list_operands timbers.
This changes list_operands to give you a list of _pointers_ to operands.
And because of const correctness, it has to be done twice (one const,
one not const). This is not a bad thing.

It does mean however that the do_type_list_operands lumps of code in
irep2.cpp become uglier.

The upside is that we can now update sub operands of exprs without any
knowledge of the expr itself. In a type safe ish manner! Of course, it'd
be stupid to update something like pointer_object2t, and change its
operand to a non-pointer. But I'd put that under the heading of "caveat
emptor".