Take another shot at base simplify method.
So, just modifying oneself is in fact a bad plan, just in case we're
reference counted. So instead, return a new copy of ourselves.

This might not actually be required... what we really want to be able to
do is have our object detatched/cloned/whatever if we're going to
simplify it, and there are other things using it. However, this is a
recipie for pain.