Realise that we can do simplify const'ly.
So, it occurs to me: in do_simplify right now, what we're doing is
creating a replacement expr and passing it upwards. We can in fact do
that for /all/ exprs, and end up with the simplification of an expr
being a new tree of exprs that's returned to the caller. This eliminates
all cares and worries about trashing reference counts.

So, constify all simplify routines. We still need non-const list_operands
to be able to manipulate the operands of an unrecognized expr
type-agnositcally.