Move large amounts of boilerplate to one method.
See previous commit; we were having to explicitly invoke
expr2t::do_simplify to simplify operands, and make a variety of path
decisions on how to do simplification, in a normal body. Clearly, copy
and pasting this is bad.

So: move it all to one method; and make do_simplifys behaviour be that
it returns a new expr if we've successfully simplified.

Now we just need the one function to attempt simplification, simplify
operands, then try simplification again. No need to put anything at all,
except attempting to simplify an expr, in the do_simplify body.