Take a shot at simplifying add2t.
This is done by moving everything to a fixedbv, operating on it, then
moving it back to what the expression type is supposed to be.

Notably, there wasn't a operator+= for fixedbvt implemented, but there
was one declared. The one I've implemented here is a vague guess which I
think will work; testing will reveal whether it does.