Make value_sett store data in irep2 format.
Once again; the plan here is to change the form of storage, and then
slowly push the envelope back upwards towards the callers.

Three new ireps are required, unknown, invalid, NULL-object. These are
going to have to hang around (and not leave the value set code!) until
it can be refactored to not use it (or I decide it's ok).