Further tweak back-conversion of constant strings.
It turns out that some code in value_sett is ensuring that everything
going into a value set has a particular type (see currently line 778
value_set.cpp), and that depends on the convertion to and from
string-constant retaining the same type. And it turns out I was using
unsignedbv where signedbv was expected.

Weird. But generally reasonable.