Constify some references.
It turns out that it's a sin to use non-C++ casts, which should have
been picked up by enabling warnings for that. Specifically, we lost the
constness on the "type" object in the code just altered, which was
leading to detatch being called, which was leading to list iterators
being invalidated. Which is bad.