Fix some type safety.
Originally before moving this for irep, this piece of code was checking
that what came out of object_numbering wasn't an "unknown" irep; wheras
I'd assumed that there would always be a dynamic_object2t coming out.
Which was wrong.

It's reassuring that this lead to a bad cast exception being thrown.