Fix type in generation of nondeterministic returns.
It turns out that this is one of the places where symbolic types can
still make their way into code. To fix this, ns.follow the type if it's
a symbol. This occurs after type-flattening, so any type symbol looked
up shouldn't contain any further symbol types.

No more or fewer cpp or normal regression tests fail as a result.