Update splicing of exprs during migration.
The problem here is the type of the expr that we calculate during
migration. If we pick pointer for some reason, or are uncareful, we end
up with exprs with multiple pointer operands, which is invalid.

So instead, for each time we splice up an expr, look at its operands and
attempt to synthesize what type it should have. Except for the top level
expr, which retains the original type.

This largely follows the rules in the C spec, 6.3.1.8.