Add a series of hacks to make --function work
So, I don't like --function. It relies on some esoteric and undefined
behaviour of CBMC in how it handles nil ireps in pointer and symex
analysis.

This is another example of it. Nil exprs make their way into the value
set analysis, with no type, and there's no good way for them to end up
becoming an "unknown" object (which is what it should become) on account
of there being no type.

So shoehorn a type into the expr when required, and at the earliest
opportunity fob the caller off with an unknown2t result.