Relax some assertions about value set op types
So, we're asserting that things like member and index should have source
operands of the correct type. However, it's perfectly valid to assign to
a dynamic object; which malloc may have decided will have a char array
type, because its own analysis failed.

So, allow index/member assignments to assign to dynamic objects of any
type.