Correct a casting situation.
We need to handle unions and structs here, not just structs.
Unfortunately because I still don't have it in me to context switch to
work on the irep implementation right now, this has to be hacked a
little rather than taking a reference to some kind of sub struct-union
type.