Move get_value_set to irep2.
Notably, this involves the first invocation of simplify. No idea if it
works yet, only add and subtract are implemented, but it doesn't crash.

One imagines as we work our way further into symex, we'll run into many
more places where the more advanced simplification is required.

More importantly though, more code is required from simplify; we
actually have to check whether or not a new anything was returned. This
should encourage sparing use of it.