Add some "code" expr ireps.
So, these should totally become something else, like a code2t base type.
However, that requires some refactoring of the whole template situation,
which I'm not wishing to do right now.

The purpose for these ireps is value_sett::apply_code, which is going to
have to handle some code expressions. In general this is a notable
point, because it's where irep starts representing code exprs, which are
the output of the parsers we have. There shouldn't be any further
expansion past this point.