Morph irep for "zero_string" to "is_zero_string".
From a brief code survey, nothign creates "zero_string" except for some
code in the ansi-c frontend that's been commented out. So, there's
precisely no point in keeping "zero_string".

And there's some code that's using "is_zero_string", which we haven't
got a conversion for, so make it use zero_string2t instead.

It's not like the z3 conversion for "zero_string" right now actually
works anyway, and is_zero_string clearly wasn't making it to the solver
anyway, because we haven't had any migration failures for it before.