Implement simplification of adds in ptr_offsets.
This algorithm copied from the old irep code. It appears to exploit the
fact that adds are only supposed to have one pointer offset. So, it can
be simplified into a pointer_offset on the pointer, which just has to be
combined with the addition.

drama: This may be where all our pointer arithmatic failure is coming
from. It's not clear whether or not pointer arith is supposed to be
handled here (with a multiply) or not.