Don't attempt to simplify past address-of's.
The problem is if we simplify its operand to some arbitary piece of
data, like an integer extracted from a constant array, we're completely
screwed as we no longer have any symbol to take the address of.

old irep code appears to attempt to simplify &a[i] to &a[0]+i, which I
suspect isn't the most effective way of doing things - I suspect that at
a later point of conversion we'd be able to take advantage of the
knowledge we're indexing an array, rather than munging with pointers.

Anyway, doesn't appear to have an particular performance impact.