Implement our associative-op-simplifying func.
The algorithm is largely as the comment at the start says; we collect
all operands of sub-exprs of the same id, then attempt to simplify any
in that list that are constant. This takes advantage of the fact that
the operation should be associative.

The return value is true if simplification occured, and presents a list
of operands for a new tree of this expr.

Incoming arguments are the expr to work, on and a function that creates
a new object of the appropriate type. I thought of templating this up,
but there are various routines that will only ever return a bool; might
want to insert a typecast; reject simplfication for some reason, or
determine a different type.