Perform simplified operations on integers directly
i.e., we use an int64_t rather than some other native type, and only
perform it if the integer being used fits in. This after the discovery
that BigInts are stored as positive internally at all times. Attempting
to work around all the operations that can be performed and what happens
with signs there is too complex. Particularly for lshr/ashr.

Most flakey part is going to be converting backwards, from int64_t to
bigint. Still haven't thought it fully through.