Correct pointer arithmetic death.
We were using the size of the array being indexed, not the size of the
elements it was made up of. This leads to wrong results because that
makes the pointer offsets all fuzzy and wrong, but more noticeably leads
to exceptions because the array isn't necessarily fixed size.

This would appear to be a success of exceptions, in letting me know when
something exceptional has occured :O

No regression tests had the wrong outcome because of this failure, which
is bad, there should be tests for these behaviours. I'm now thinking
there should be a regression tests for every line of code in here :|