Increase accuracy of casting via fixedbvs.
That is: if we cast a fixedbv to a smaller integer type, we want to drop
higher bits. We also *must* generate an inverted bit representation when
casting from negative to unsigned, or the world explodes.

One imagines random death occurs in integer mode; but it would, because
it's built on top of a codebase that didn't support it in the first
place :|