Attempt to fix integer bit truncations in fixedbvt
That is, if we want to round off the /top/ of a fixedbvt, do so via
modulusing. This is required to make fixedbvts truncate to the right
lengths, so that they can be converted back to exprs with expected
types. I.e. 64/32 and 32/16 widths. If they don't, simplification in old
irep fails.

This requires patching the esbmc makefiles because the "pow" function is
in another arbitary binary, and I don't have the will to significantly
edit a bunch of makefiles right now.