Normalise fixedbv widths before operating on them.
There doesn't appear to be any kind of check in the fixedbvt type that
when we do two operations with fixedbvts, that they have the same
integer representation internally. Work around this by ensuring that
they always have the same bit fractions / etc.

This just scales up anything to a higher accuracy, if needed.